Potentially fixed Protobuf reset issue

+++ main/main.c
@@ -85,14 +85,15 @@ void master_task(void *pvParameter){
                 // update
                 robotState.inBallAngle = floatMod(lastSensorUpdate.tsopAngle + TSOP_CORRECTION, 360.0f);
                 robotState.inBallStrength = lastSensorUpdate.tsopStrength;
+                // ESP_LOGD(TAG, "Values of SensorUpdate: heading: %f, ball angle: %f, ball strength: %f",
+                // lastSensorUpdate.heading, lastSensorUpdate.tsopAngle, lastSensorUpdate.tsopStrength);
 
                 robotState.inGoalVisible = robotState.outIsAttack ? AWAY_GOAL.exists : HOME_GOAL.exists;
                 robotState.inGoalAngle = robotState.outIsAttack ? AWAY_GOAL.angle + CAM_ANGLE_OFFSET : HOME_GOAL.angle 
                                         + CAM_ANGLE_OFFSET;
                 // TODO make goal angle a float as well
-                robotState.inGoalLength = robotState.outIsAttack ? (int16_t) AWAY_GOAL.length : HOME_GOAL.length;
+                robotState.inGoalLength = robotState.outIsAttack ? (int16_t) AWAY_GOAL.length : (int16_t) HOME_GOAL.length;
                 robotState.inGoalDistance = robotState.outIsAttack ? AWAY_GOAL.distance : HOME_GOAL.distance;
-                // hack to convert to IMU data to float by multiplying it by 100 before sending then diving it
                 robotState.inHeading = lastSensorUpdate.heading;
                 robotState.inX = robotX;
                 robotState.inY = robotY;
@@ -108,11 +109,6 @@ void master_task(void *pvParameter){
         // update the actual FSM
         fsm_update(fsm);
         // ESP_LOGI(TAG, "State: %s", fsm_get_current_state_name(fsm));
-
-        // update_line(&robotState);
-
-        // robotState.outSpeed = 0;
-        // print_ball_data(&robotState);
 
         // run motors
         motor_calc(robotState.outDirection, robotState.outOrientation, robotState.outSpeed); // Our silly old motor code
